requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
certifi>=2024.0.0
django-tidb>=4.2.0
mysqlclient>=2.2.0,<3.0.0
django-cors-headers>=4.3,<5.0

# Production dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
